requirements.txt
python-dotenv>=1.0.0
jinja2>=3.1.0
requests>=2.31.0
python-dateutil>=2.8.0
feedparser>=6.0.0
openai>=1.0.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1 